+++ CHANGELOG.md
@@ -6,6 +6,8 @@
   - matrices: support setting `split` property as a string to allow using children's indexes in their address (`$` is replaced with the index number)
   - text: `wrap` property doesn't break words by default anymore
   - rgb: added `range` property
+- launcher
+  - right-click context menu (copy, paste, clear console, toggle console autoscoll)
 
 ## 0.27.2